pkc-firmware: PKC firmware running in EP side
C29XPCIE cards are a series of Freescale PCIe add-in cards working on
EP side to perform public key crypto acceleration. Here is the diagram
for PKC hardware setup and software components.

          RC (X86-64/PowerPC)            EP (C293PCIE card)
	-----------------------        -----------------------
	|   OpenSSL           |        |                     |
	|----------------------        |                     |
	|   CryptoDev         |        |   pkc-firmware      |
	|   CryptoAPI         |        |     (U-Boot)        |
	|   pkc-host driver   |        |                     |
	|                     |	 PCIe  |                     |
	|   PCIe data path    |<======>|   PCIe data path    |
	-----------------------        -----------------------

The pkc-firmware is the driver running in the EP side to push the
encryption/decryption operations from host to SEC engine and return
the results to host. pkc-host driver is maintained in another tree.

Signed-off-by: Roy Zang <tie-fei.zang@freescale.com>
Signed-off-by: Yuanquan Chen <Yuanquan.Chen@freescale.com>
Signed-off-by: Mingkai Hu <Mingkai.Hu@freescale.com>