remove unnecessary commented instruction
The "SEC_ENG_COUNT" define was used to statically initialize
the "sec_eng_count" variable. A better way has been implemented by
checking the c29x registers and using the value found.
There is no more trace in the code of the "SEC_ENG_COUNT" define (except
the one in the commented instruction) suggesting that the latter method
of initialization was preferred.

Signed-off-by: Radu Alexe <raduandrei.alexe@freescale.com>