enable large bursts for DECO transactions
Enabling LARGE_BURST increases transaction size and thus either reduces
the number of required transactions or permits more data throughput with
the same transaction count. This effectively improves the CAAM performance
but does not directly translate in performance improvements for all cases
(the generation of larger bursts from DECOs depend on how fast FIFOs fill
and drain).

This feature is present in C293 but not in earlier versions of CAAM and for
reasons of forward and backward compatibility (not changing default
documented register values) it is not enabled by default.

From C293 Reference manual:
"When LARGE_BURST=1, Job Descriptor reads, Shared Descriptor reads and
reads of data for the Data FIFO can use transactions as large as the
maximum AXI interface transaction LARGE_BURST size, which equals 16
times the width of the AXI data buses (i.e. 128 bytes for 64-bit data
buses)."

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>