rename variable "p_cursor" and make it global
This variable is used throughout the firmware and it is always passed as
a function parameter. A better approach should be to make it global and
eliminate the need to always pass it as an argument.

This is the first of a series of changes that strives to isolate the
implementation regarding memory allocation. The chief end is to provide
the firmware code with a "malloc/kalloc" like function that should provide
an abstraction regarding the memory management process.

This strategy provide simplicity, familiarity and clarity for the memory
partition/allocation.

Signed-off-by: Radu Alexe <raduandrei.alexe@freescale.com>