simplify calculation of address alignment
We leave out only one alignment function since all the use cases
can be solved with it. In fact there is only one use-case where the
alignment is necessary: when a data structure must not share a cache
line with adjacent structures. To do this, it is enough to align the
stack pointer before and after allocation. Since the alignment will
return a smaller address this is consistent with the fact that the
stack pointer grows downward.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>