fix interrupt notification to Intel(R) hosts with IOMMU enabled
This patch fixes an error codition on Intel(R) platforms with
interrupt-remapping support.

Linux Kernel reports this error as:
dmar: INTR-REMAP: Request device [[01:00.0] fault index 1e
INTR-REMAP:[fault reason 32] Detected reserved fields in the decoded interrupt-remapped request

The PCI spec for MSI is not exactly clear about this but hindsight is
20-20. Interrupt requests are DOUBLE WORDS = 32 bits:

- Message Data register (16 bits) must be written to AD[15::00] of the
  interrupt address
- In the same memory write, bits AD[31::16] must be set to zero.

If the upper bits are not set to zero, things work because these bits
are (seem to be) ignored. On Intel(R) platforms with IOMMU is enabled,
this is no longer the case because Interrupt Requests are in Remappable
Format (confirmed by bit SHV of the Interrupt Address) and the upper 16
bits _must_ be set to zero or a failure mode occurs.

See also:
Intel(R) Virtualization Technology for Directed I/O Architecture Specification, Rev 2.3
- 5.1.2 Interrupt Requests Format on Intel(R) 64 Platforms

PCI LOCAL BUS SPECIFICATION, Rev. 3.0
- 6.8.1.6 Message Data for MSI

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>