make explicit that we have a single response ring hardcoded to id 0
This was always the case but with the confusion comming from
mixing drv_response_ring and fw_response_ring, and that some parts of
the code assumed more than one response rings there was not much fun in
looking at this code.

Hard code the response ring to id 0 to set up path for true multiple
response rings.

In the process, the structure drv_resp_ring_t turns out not to be
necessary and we remove it.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>