make explicit that we are using C89/90 standard with GNU extensions
This should be the same as the standard we're using in the host driver.
Even if the u-boot code requires a different standard than GNU C90, the
pkc-firmware itself standard is the one we set here.

This is a preemtive update in expectation of the GCC 5.0 toolchain
updates in NXP SDK which by default is using C11 standard.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>