remove the assumption that ring 0 has any special significance
Ring 0 used to be the "command" ring. We've removed this specialization
and this patch follows up removing something that we missed.
The rng code is wrong in both comment and code. The rng init used the
ring 0 code (not 1) but 0 was for commands not for data.
Again, this strikes as an undetected abstraction leak.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>