simplify calculation of msi address
Save the value of MSI address given by the host and reuse it to set tlb.
We could get the same MSI address by reading locally from the PCIe
registers - but we save this work for another day.
Since we're using MSI, there is a single address for all vectors.
Instead of communicating it to the firmware for each ring plus one more
time before intialization, use a single point of truth only the rings data
as this would have been the appropriate thing to do with MSIx where we
have different addresses for each vector.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>