ring affinity: keep jobs and responses in the same ring pair
This functionality aims at improving data locality on the host. The
descriptor will be returned to the same ring pair used for submission.
This patch maintains only ring affinity for the jobs. This is a
necessary but not sufficient step to maintain CPU affinity.

OBS: The CPU used on the host for sending jobs is not necessarily the
same as the one used by the IRQ handler so CPU affinity is not
maintained across all the host-device loop.
This patch stretches the E500 capabilities. We've slighly lower
maximum performance on the RSA1024 public operations but the host CPU
load is also lower due to improved data locality. This is confirmed by
the results with RSA2048 which do not stress the E500 core at all.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>